requirements.txt
# Requirements for LinkedIn Blog AI Assistant

# Core LLM and framework dependencies
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-groq
langchain-community
langgraph

# Google Generative AI (required for image_pipeline.py)
google-generativeai

# FastAPI web framework
fastapi
uvicorn[standard]
python-multipart

# Image processing and computer vision
Pillow
opencv-python
pdf2image

# PDF processing
PyPDF2
pymupdf  # fitz
pypdf
# Presentation processing
python-pptx

# Web and API utilities
requests
beautifulsoup4

# Data handling
pandas
numpy

# Environment and utilities
python-dotenv
typing-extensions

tavily-python

# DeepAgents for advanced reasoning
deepagents